+++ src/ServiceContracts/Tenants/TemplatesRawContract.php
@@ -8,8 +8,12 @@ use Courier\Core\Contracts\BaseResponse;
 use Courier\Core\Exceptions\APIException;
 use Courier\RequestOptions;
 use Courier\Tenants\BaseTemplateTenantAssociation;
+use Courier\Tenants\PostTenantTemplatePublishResponse;
+use Courier\Tenants\PutTenantTemplateResponse;
 use Courier\Tenants\Templates\TemplateListParams;
 use Courier\Tenants\Templates\TemplateListResponse;
+use Courier\Tenants\Templates\TemplatePublishParams;
+use Courier\Tenants\Templates\TemplateReplaceParams;
 use Courier\Tenants\Templates\TemplateRetrieveParams;
 
 /**
@@ -48,6 +52,40 @@ interface TemplatesRawContract
     public function list(
         string $tenantID,
         array|TemplateListParams $params,
+        RequestOptions|array|null $requestOptions = null,
+    ): BaseResponse;
+
+    /**
+     * @api
+     *
+     * @param string $templateID path param: Id of the template to be published
+     * @param array<string,mixed>|TemplatePublishParams $params
+     * @param RequestOpts|null $requestOptions
+     *
+     * @return BaseResponse<PostTenantTemplatePublishResponse>
+     *
+     * @throws APIException
+     */
+    public function publish(
+        string $templateID,
+        array|TemplatePublishParams $params,
+        RequestOptions|array|null $requestOptions = null,
+    ): BaseResponse;
+
+    /**
+     * @api
+     *
+     * @param string $templateID path param: Id of the template to be created or updated
+     * @param array<string,mixed>|TemplateReplaceParams $params
+     * @param RequestOpts|null $requestOptions
+     *
+     * @return BaseResponse<PutTenantTemplateResponse>
+     *
+     * @throws APIException
+     */
+    public function replace(
+        string $templateID,
+        array|TemplateReplaceParams $params,
         RequestOptions|array|null $requestOptions = null,
     ): BaseResponse;
 }