+++ src/Tenants/TenantTemplateInput/Channel/RoutingMethod.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Courier\Tenants\TenantTemplateInput\Channel;
+
+/**
+ * Defaults to `single`.
+ */
+enum RoutingMethod: string
+{
+    case ALL = 'all';
+
+    case SINGLE = 'single';
+}