+++ tests/Services/Users/PreferencesTest.php
@@ -4,6 +4,7 @@ namespace Tests\Services\Users;
 
 use Courier\ChannelClassification;
 use Courier\Client;
+use Courier\Core\Util;
 use Courier\PreferenceStatus;
 use Courier\Users\Preferences\PreferenceGetResponse;
 use Courier\Users\Preferences\PreferenceGetTopicResponse;
@@ -25,7 +26,7 @@ final class PreferencesTest extends TestCase
     {
         parent::setUp();
 
-        $testUrl = getenv('TEST_API_BASE_URL') ?: 'http://127.0.0.1:4010';
+        $testUrl = Util::getenv('TEST_API_BASE_URL') ?: 'http://127.0.0.1:4010';
         $client = new Client(apiKey: 'My API Key', baseUrl: $testUrl);
 
         $this->client = $client;