+++ tests/Services/Users/TenantsTest.php
@@ -3,6 +3,7 @@
 namespace Tests\Services\Users;
 
 use Courier\Client;
+use Courier\Core\Util;
 use Courier\Users\Tenants\TenantListResponse;
 use PHPUnit\Framework\Attributes\CoversNothing;
 use PHPUnit\Framework\Attributes\Test;
@@ -21,7 +22,7 @@ final class TenantsTest extends TestCase
     {
         parent::setUp();
 
-        $testUrl = getenv('TEST_API_BASE_URL') ?: 'http://127.0.0.1:4010';
+        $testUrl = Util::getenv('TEST_API_BASE_URL') ?: 'http://127.0.0.1:4010';
         $client = new Client(apiKey: 'My API Key', baseUrl: $testUrl);
 
         $this->client = $client;